+++ CHANGELOG.md
@@ -3,9 +3,12 @@
 ## 1.0.0-RC.3 - Unreleased
 
 - The package now includes Datastar [1.0.0-RC.3](https://github.com/starfederation/datastar/releases/tag/v1.0.0-RC.3).
-- Added the ability to pass controller actions to backend requests by passing an array `['MyController', 'myAction']` to `datastar()->get()` (and equivalent Blade directives).
-- Added the ability to pass routes to backend requests by passing a URI that begins with `'http` or `/` to `datastar()->get()` (and equivalent Blade directives).
-- Server sent events are now only sent when generated inside a callable passed to `getEventStream()`.
+- The `datastar()->get()` and equivalent Blade directives now only support passing a route URI.
+- Added the ability to render views in backend requests by passing a view path `path.to.view` to `datastar()->view()`.
+- Added the ability to pass controller actions to backend requests by passing an array `['MyController', 'myAction']` to `datastar()->action()`.
+- Replaced the `DatastarEventStream` trait with the `sse()` helper.
+- Added the `getValidator()` method to the `sse()` helper.
+- Added the `validate()` and `validateWithBag()` methods to the `sse()` helper.
 - Renamed the `getStreamedResponse()` method to `getEventStream()`.
 - Renamed the `renderDatastarView()` method to `renderView()`.